requirements/production.txt
django==1.9.7
dj-database-url
psycopg2
djangorestframework
djangorestframework-jwt
Pillow
honcho
gunicorn
geopy
requests
django-storages-redux
boto
django-pipeline
raven
